+++ test/unit/lib/standard.specs.js
@@ -14,7 +14,7 @@ test.describe('standard', () => {
     let mocksSandbox
 
     test.beforeEach(() => {
-      sandbox = test.sinon.sandbox.create()
+      sandbox = test.sinon.createSandbox()
       mocksSandbox = new mocks.Sandbox([
         'logs',
         'paths',