+++ test/unit/lib/suite-docker.specs.js
@@ -14,7 +14,7 @@ test.describe('suite-docker', () => {
   let mocksSandbox
 
   test.beforeEach(() => {
-    sandbox = test.sinon.sandbox.create()
+    sandbox = test.sinon.createSandbox()
     mocksSandbox = new mocks.Sandbox([
       'docker',
       'paths',